% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griffithlab-package.R
\docType{package}
\name{griffithlab-package}
\alias{griffithlab}
\alias{griffithlab-package}
\title{griffithlab: Tools for the lab of Jamie Griffith}
\description{
Assorted analysis tools.
    Scoring algorithms, etc.
}
\section{Words of inspiration}{

Le bon dieu est dans le détail, Gustave Flaubert (1821-1880)
}

\keyword{internal}
