% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_pfdi.R
\name{score_pfdi}
\alias{score_pfdi}
\title{Returns scores for the 20-item Pelvic Floor Distress Inventory
(PFDI) based on a dataframe of input.}
\usage{
score_pfdi(input, transfer_vars = names(input))
}
\arguments{
\item{input}{A dataframe containing PFDI items. Other columns may also
be present and will be returned by the function (if desired). The items must
use the recommended names: pfdi_20_1, pfdi_20_1a, pfdi_20_2, pfdi_20_2a.
Factor variables will result in an error.}

\item{transfer_vars}{A vector of variable names to be found in input.
These variables will be returned in the output along with the
PFDI scores.}
}
\value{
A dataframe of output containing PFDI scores.
}
\description{
This function returns takes a dataframe, extracts the
PFDI-20 items, calculates subscale scores, and returns the
scores along with any other requested variables from the input.
}
\details{
If only a subset of variables are desired to be returned,
the column names can be specified in transfer_vars.
It should be noted that if any out-of-range or non-numeric
(e.g., character) values are detected, they are re-coded to NA.
}
\section{Item response coding}{
 Items are coded 0/1 for the No/Yes items and
Items "a" are coded 1-4. This coding must be respected.
}

\examples{
\dontrun{
score_pfdi(input = some_data)
}
}
