% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_skindex_teen.R
\name{score_skindex_teen}
\alias{score_skindex_teen}
\title{Scores the Skindex-Teen questionnaire using only the 21 scored items
from the Smidt et al. paper (cf., Table 1 of doi:10.1001/archdermatol.2010.161)}
\usage{
score_skindex_teen(items, min_prop = 1, skindex_teen_names = NULL)
}
\arguments{
\item{items}{A matrix (or an object coercible to a matrix) that contains
the items of the Skindex-teen, with each item
represented as a number, 0, 1, 2, 3, or 4. Anything not a 0, 1, 2, 3, or 4 will
silently be converted to NA.}

\item{min_prop}{The minimum proportion of items needed to be non-missing
in order for a score to be given. If the observed proportion of non-missing items is
less than min_prop, then the score will be NA. Otherwise, in the
presence of missing data, prorating will be used. With prorating, the score
is calculated as (# of scale items * mean item response) The default for this
algorithm is min_prop = 1.0.}

\item{skindex_teen_names}{If left NULL, by default names in the output
will be "skindex_teen_total", "skindex_teen_ps", and "skindex_teen_pf"}
}
\value{
Scores for the Skindex-Teen Total,
Physical Symptoms (ps) and Psychosocial Functioning (pf).
}
\description{
Calculates a total score of the 21 Skindex-Teen items
(i.e., skindex_teen_1 through skindex_teen_21), as well as two subscales
for Physical Symptoms (ps, 5 items) and Psychosocial Functioning (pf, 16 items).
If a vector is supplied as input, it will be converted to a one-row,
21-column dataframe. Items are assumed to be in the same order as in Table 1
of Smidt et al. but with item 17 of 22 removed.
}
\examples{
\dontrun{
score_skindex_teen(data[skindex_teen_items])
}
}
