% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griffithlab_internal_functions.R
\name{score_surveys}
\alias{score_surveys}
\title{Internal generic function for helping to score surveys}
\usage{
score_surveys(items, min_num_items = ncol(items))
}
\arguments{
\item{items}{A dataframe of numeric  questionnaire items.}

\item{min_num_items}{The minimum number of items needed to score the
questionnaire. If not enough items are present, the score will be NA.
If some items are missing, but the number of non-missing items is
min_num_items or higher, then the score will be prorated.}
}
\value{
Scores for each case (i.e., row)
}
\description{
Takes a dataframe of numeric items and calculates a score
}
\examples{
# score_surveys(rrs, 10)
# score_surveys(rrs, 8)
}
\keyword{internal}
