% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_sus.R
\name{score_sus}
\alias{score_sus}
\title{Scores the System Usability Scale (SUS; all 10 items)}
\usage{
score_sus(sus_items, min_num_items = 10)
}
\arguments{
\item{sus_items}{A matrix (or an object coercible to a matrix) that contains
the items of the System Usability Scale (SUS), with each item
represented as a number, 1, 2, 3, 4, or 5.}

\item{min_num_items}{The minimum number of items needed to be non-missing
in order for a score to be given. If the number of non-missing items is
less than min_num_items, then the score will be NA. Otherwise, in the
presence of missing data, prorating will be used. With prorating the score
is (10 * mean item response) The default for the SUS is min_num_items = 10.}
}
\value{
Scores for the SUS.
}
\description{
Scores the 10-item System Usability Scale (SUS)
The ten items are 1, 2, 3, 4, 5 coded in the following way:\cr
1 = Strongly Disagree\cr
2 = Disagree\cr
3 = Neutral\cr
4 = Agree\cr
5 = Strongly Agree\cr
}
\details{
In some versions of the SUS, only 1 (Strongly Disagree) and 5 (Strongly Agree) are labelled.
}
\examples{
\dontrun{
sus_items <- paste0("sus", 1:10)
score_sus(some_data[sus_items])
}
}
