% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_letter_codes.R
\name{two_letter_codes}
\alias{two_letter_codes}
\title{Return two-letter codes with unique letters and no reversals}
\usage{
two_letter_codes(x = LETTERS)
}
\arguments{
\item{x}{Set to LETTERS by default ("A", "B", "C", and so on.)}
}
\value{
All possible two-letter codes with no repeats and with no reversals.
}
\description{
This function returns two-letter codes.
For each value, the letters will be unique (e.g., "AB", but never "AA").
Also, there will be no reversals (e.g., "AB", but never "BA").
This function is used to facilitate randomisation codes for experiments and
clinical trials
}
\details{
Only uppercase letters are used. If desired, the user can delete letters
that resemble numerals, and  similar-looking letters,
e.g., I, L, O, U, V, and Z (and maybe others).
The input x will sorted in ascending order.
}
\examples{
\dontrun{
two_letter_codes()
two_letter_codes(LETTERS[1:10]
two_letter_codes("A", "B", "C", "D", "E", "F")
}
}
